Torch sim update
